server/requirements.txt
fastapi
uvicorn
jinja2
pydantic-settings
google-generativeai
python-dotenv
